Assessment-Index-Modelling-master/requirements.txt
python==3.9.13
pandas==1.4.4
python-dateutil==2.8.2